made compatible with zvtm 0.11.0
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3774 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm-nodetrix/trunk/src/main/java/fr/inria/zvtm/glyphs/GPath.java
@@ -34,7 +34,7 @@ public class GPath extends DPath {
 		*@param z z-index (pass 0 if you do not use z-ordering)
 		*@param c color
 		*/
-	public GPath(long x, long y, int z, Color c){
+	public GPath(double x, double y, int z, Color c){
 	    super(x, y, z, c);
     }
     
@@ -45,7 +45,7 @@ public class GPath extends DPath {
 		*@param c color
 		*@param alpha alpha channel value in [0;1.0] 0 is fully transparent, 1 is opaque
 		*/
-	public GPath(long x, long y, int z, Color c, float alpha){
+	public GPath(double x, double y, int z, Color c, float alpha){
 		super(x, y, z, c, alpha);
 	}
